+++ crates/sui-indexer/src/models_v2/epoch.rs
@@ -4,7 +4,7 @@
 use diesel::{Insertable, Queryable, Selectable};
 
 use crate::errors::IndexerError;
-use crate::schema_v2::epochs;
+use crate::schema::epochs;
 use crate::types_v2::IndexedEpochInfo;
 use sui_json_rpc_types::{EndOfEpochInfo, EpochInfo};
 use sui_types::sui_system_state::sui_system_state_summary::SuiSystemStateSummary;
