+++ crates/sui-indexer/src/models_v2/objects.rs
@@ -15,7 +15,7 @@ use sui_types::object::Object;
 use sui_types::object::ObjectRead;
 
 use crate::errors::IndexerError;
-use crate::schema_v2::{objects, objects_history};
+use crate::schema::{objects, objects_history};
 use crate::types_v2::{IndexedDeletedObject, IndexedObject, ObjectStatus};
 
 #[derive(Queryable)]
