+++ crates/sui-indexer/src/models_v2/transactions.rs
@@ -17,7 +17,7 @@ use sui_types::event::Event;
 use sui_types::transaction::SenderSignedData;
 
 use crate::errors::IndexerError;
-use crate::schema_v2::transactions;
+use crate::schema::transactions;
 use crate::types_v2::IndexedObjectChange;
 use crate::types_v2::IndexedTransaction;
 use crate::types_v2::IndexerResult;
