--- crates/sui-indexer/src/schema_v2.patch
@@ -1,7 +1,0 @@
-diff --git a/crates/sui-indexer/src/schema.rs b/crates/sui-indexer/src/schema.rs
---- a/crates/sui-indexer/src/schema.rs
-+++ b/crates/sui-indexer/src/schema.rs
-@@ -1 +1,3 @@
-+// Copyright (c) Mysten Labs, Inc.
-+// SPDX-License-Identifier: Apache-2.0
- // @generated automatically by Diesel CLI.
