+++ crates/sui-indexer/src/store/module_resolver_v2.rs
@@ -19,7 +19,7 @@ use crate::errors::{Context, IndexerError};
 use crate::handlers::tx_processor::IndexingPackageBuffer;
 use crate::metrics::IndexerMetrics;
 use crate::models_v2::packages::StoredPackage;
-use crate::schema_v2::{objects, packages};
+use crate::schema::{objects, packages};
 use crate::store::diesel_macro::read_only_blocking;
 use crate::types_v2::IndexedPackage;
 use crate::PgConnectionPool;
