+++ crates/sui-indexer/src/store/pg_indexer_store_v2.rs
@@ -38,7 +38,7 @@ use crate::models_v2::objects::{
 };
 use crate::models_v2::packages::StoredPackage;
 use crate::models_v2::transactions::StoredTransaction;
-use crate::schema_v2::{
+use crate::schema::{
     checkpoints, display, epochs, events, objects, objects_history, objects_snapshot, packages,
     transactions, tx_calls, tx_changed_objects, tx_input_objects, tx_recipients, tx_senders,
 };
